Rename TileInfo.h to TileMapping.h
